Made Vector2's operator overloading return a type depending on the result of the operation Instead of depending on the first member, units also don't need that much precision, there's not much difference, nor bad accuracy

+++ include/Datatypes/Units.h
@@ -1,4 +1,4 @@
 #pragma once
 
 // Usually a floating point datatype that is used to describe coordinates
-typedef double Unit;
+typedef float Unit;
